Polish kalman_deneme.c
The predicted array is updated with respect to the previous actual array.

Some new variables are introduced for the sake of pseudoinverse calculation:
arrays a and b, float c
These should probably be renamed. (horrible names :D)

The determinant became 0 after one or two iterations, hence a pseudoinverse is used now, and it seems to work.

A delay of 500 ms is added.